changed HomeRoutes code
