--- public/js/genre.js
@@ -1,51 +1,0 @@
-$("#genre-list button").on("click", function(event) {
-  const genre = event.target.textContent;
-  console.log("getting books for " + genre);
-
-  fetch(`/api/books/${genre}`)
-    .then(res => res.json())
-    .then(data => {
-      console.log(data);
-      $("#results-list").empty(); // Clear previous results
-
-      data.forEach(book => {
-        const newDiv = $("<div>");
-        newDiv.addClass("card");
-        newDiv.css("width", "18rem");
-
-        newDiv.html(`
-          <div class="card">
-            <img class="card-img-top" src="${book.volumeInfo.imageLinks ? book.volumeInfo.imageLinks.smallThumbnail : 'https://via.placeholder.com/128x192'}" alt="Book Cover">
-            <div class="card-body">
-              <h5 class="card-title">${book.volumeInfo.title}</h5>
-              <p class="card-text">${book.volumeInfo.description || 'No description available'}</p>
-              <button class="btn btn-primary add-to-favorites" data-book-id="${book.id}">Add to favorites</button>
-              </div>
-            </div>
-          `);
-  
-          $("#results-list").append(newDiv);
-        });
-      });
-  });
-  
-  
-$("#results-list").on("click", ".add-to-favorites", function(event) {
-  const bookId = $(this).data("book-id");
-  const bookTitle = $(this).data("book-title"); 
-  addToFav(bookId, bookTitle);
-});
-
-const addToFav = async (bookId, bookTitle) => {
-  const response = await fetch(`/api/favorites`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ book_id: bookId }),
-  });
-
-  if (response.ok) {
-    console.log(`${bookTitle} has been added to favorites.`);
-alert(`${bookTitle} has been added to favorites.`);
-  }
-};
-
