+++ server.js
@@ -2,6 +2,7 @@ const path = require("path");
 const express = require('express');
 const session = require('express-session');
 //const axios = require("axios");
+const db = require('./models');
 const exphbs = require('express-handlebars');
 const routes = require('./controllers');
 const sequelize = require('./config/connection');
