+++ views/login.handlebars
@@ -55,4 +55,4 @@
           </section>
         
         
-        <script src="/js/login.js"></script>
+        <script src="./js/login.js"></script>
